Extract shared field resolvers into resolve_fields.go
Move the field-by-name and field-by-ID resolvers (ResolveFieldByName,
ResolveFieldByID) and their error types (FieldNotFoundError,
FieldAmbiguousError, FieldIDNotFoundError) out of resolve.go into a new,
self-contained resolve_fields.go, with the matching tests in
resolve_fields_test.go.

These resolvers are the only part of the shared name-resolution layer that
the item-list read path also needs. Keeping them in their own file lets the
item-list PR ship that same file byte-identically, so the two PRs can merge
independently in any order (Git auto-merges an identical "added by both"
file into a single copy) instead of forcing a stacked base.

No behavior change; resolve.go keeps the option and item resolvers.
